TrickyState.wanderTargetY: u32 -> f32 (every consumer wrote it through a *(f32*)& pun; matches wanderTargetX/Z)

+++ include/main/dll/tricky_state.h
@@ -235,7 +235,7 @@ typedef struct TrickyState {
     u8 stateFlags728; /* flag byte: 0/1 boolean sets (tricky/animobjd2) plus a bit-5 test (tricky_substates) */
     u8 pad729[0x72C - 0x729];
     f32 wanderTargetX; /* wander/return target position X that targetPosPtr is pointed at (&wanderTargetX); written from anim world/local posX plus a sin offset (tricky/tricky_substates) */
-    u32 wanderTargetY; /* wander target position Y (written as f32 from anim world/local posY) */
+    f32 wanderTargetY; /* wander target position Y (written from anim world/local posY) */
     f32 wanderTargetZ; /* wander target position Z: anim world/local posZ plus a cos offset; a rare u8-overlay at this offset toggles a state bit (tricky/tricky_substates) */
     f32 sfxRepeatTimer; /* f32 countdown: -= timeDelta, on reaching floor fires an SFX and re-primes to lbl_803E2440 (tricky_substates) */
     f32 moveHoldTimer; /* f32 countdown primed to randomGetRange(120,240) on entering idle move 0x29; counted down in move 0x2a and on reaching the floor advances to move 0x2b or 0x2c (tricky_substates) */
